# N3 store perf experiment