--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -130,6 +130,79 @@ N3Store.prototype = {
     return results;
   },
 
+  _someInIndex: function (index0, key0, key1, key2, name0, name1, name2, graph, fn) {
+    var tmp, index1, index2, varCount = !key0 + !key1 + !key2,
+        // depending on the number of variables, keys or reverse index are faster
+        entityKeys = varCount > 1 ? Object.keys(this._ids) : this._entities;
+
+    // If a key is specified, use only that part of index 0.
+    if (key0) (tmp = index0, index0 = {})[key0] = tmp[key0];
+    for (var value0 in index0) {
+      var entity0 = entityKeys[value0];
+
+      if (index1 = index0[value0]) {
+        // If a key is specified, use only that part of index 1.
+        if (key1) (tmp = index1, index1 = {})[key1] = tmp[key1];
+        for (var value1 in index1) {
+          var entity1 = entityKeys[value1];
+
+          if (index2 = index1[value1]) {
+            // If a key is specified, use only that part of index 2, if it exists.
+            var values = key2 ? (key2 in index2 ? [key2] : []) : Object.keys(index2);
+            // Create triples for all items found in index 2.
+            for (var l = values.length - 1; l >= 0; l--) {
+              var result = { subject: '', predicate: '', object: '', graph: graph };
+              result[name0] = entity0;
+              result[name1] = entity1;
+              result[name2] = entityKeys[values[l]];
+              if (fn(result.subject, result.predicate, result.object, result.graph)) {
+                return true;
+              }
+            }
+          }
+        }
+      }
+    }
+    return false;
+  },
+
+  _lookupIndex0lookupIndex1loopIndex2: function (index0, key0, key1, fn) {
+    var index1, index2, key2;
+    // Lookup.
+    if (!(index1 = index0[key0])) return;
+    // Lookup.
+    if (!(index2 = index1[key1])) return;
+    // Loop.
+    for (key2 in index2)
+      fn(key2);
+  },
+
+  _lookupIndex0loopIndex1: function (index0, key0, fn) {
+    var index1, key1;
+    // Lookup.
+    if (!(index1 = index0[key0])) return;
+    // Loop.
+    for (key1 in index1)
+      fn(key1);
+  },
+
+  _loopIndex0lookupIndex1: function (index0, key1, fn) {
+    var key0, index1;
+    // Loop.
+    for (key0 in index0) {
+      index1 = index0[key0];
+      // Lookup.
+      if (index1[key1])
+        fn(key0);
+    }
+  },
+
+  _loopIndex0: function (index0, fn) {
+    // Loop.
+    for (var key0 in index0)
+      fn(key0);
+  },
+
   // ### `_countInIndex` counts matching triples in a three-layered index.
   // The index base is `index0` and the keys at each level are `key0`, `key1`, and `key2`.
   // Any of these keys can be `null`, which is interpreted as a wildcard.
@@ -361,6 +434,382 @@ N3Store.prototype = {
       return this._countInIndex(graphItem.objects, object, subject, predicate);
     }
   },
+
+  forEach: function (fn, subject, predicate, object, graph) {
+    var prefixes = this._prefixes;
+    this.forEachByIRI(
+      fn,
+      expandPrefixedName(subject,   prefixes),
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(object,    prefixes),
+      expandPrefixedName(graph,     prefixes)
+    );
+  },
+
+  forEachByIRI: function (fn, subject, predicate, object, graph) {
+    this.someByIRI(function (s, p, o, g) {
+      fn(s, p, o, g);
+      return false;
+    }, subject, predicate, object, graph);
+  },
+
+  every: function (fn, subject, predicate, object, graph) {
+    var prefixes = this._prefixes;
+    return this.everyByIRI(
+      fn,
+      expandPrefixedName(subject,   prefixes),
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(object,    prefixes),
+      expandPrefixedName(graph,     prefixes)
+    );
+  },
+
+  everyByIRI: function (fn, subject, predicate, object, graph) {
+    return !this.someByIRI(function (s, p, o, g) {
+      return !fn(s, p, o, g);
+    }, subject, predicate, object, graph);
+  },
+
+  some: function (fn, subject, predicate, object, graph) {
+    var prefixes = this._prefixes;
+    return this.someByIRI(
+      fn,
+      expandPrefixedName(subject,   prefixes),
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(object,    prefixes),
+      expandPrefixedName(graph,     prefixes)
+    );
+  },
+
+
+  someByIRI: function (fn, subject, predicate, object, graph) {
+    var graphs = {}, graphContents,
+        ids = this._ids, subjectId, predicateId, objectId;
+    // Either loop over all graphs, or over just one selected graph.
+    if (!graph)
+      graphs = this._graphs;
+    else
+      graphs[graph] = this._graphs[graph];
+
+    // Translate IRIs to internal index keys.
+    // Optimization: if the entity doesn't exist, no triples with it exist.
+    if (subject   && !(subjectId   = ids[subject]))   return false;
+    if (predicate && !(predicateId = ids[predicate])) return false;
+    if (object    && !(objectId    = ids[object]))    return false;
+
+    for (var graphId in graphs) {
+      // Only if the specified graph contains triples, there can be results
+      if (graphContents = graphs[graphId]) {
+        // Do not emit the default graph explicitly
+        if (graphId === DEFAULT_GRAPH)
+          graphId = '';
+
+        // Choose the optimal index, based on what fields are present
+        if (subjectId) {
+          if (objectId) {
+          // If subject and object are given, the object index will be the fastest.
+            if (this._someInIndex(graphContents.objects, objectId, subjectId, predicateId,
+                                  'object', 'subject', 'predicate', graphId, fn))
+              return true;
+          }
+          else
+            // If only subject and possibly predicate are given, the subject index will be the fastest.
+            if (this._someInIndex(graphContents.subjects, subjectId, predicateId, null,
+                                  'subject', 'predicate', 'object', graphId, fn))
+              return true;
+        }
+        else if (predicateId) {
+          // If only predicate and possibly object are given, the predicate index will be the fastest.
+          if (this._someInIndex(graphContents.predicates, predicateId, objectId, null,
+                                'predicate', 'object', 'subject', graphId, fn)) {
+            return true;
+          }
+        }
+        else if (objectId) {
+          // If only object is given, the object index will be the fastest.
+          if (this._someInIndex(graphContents.objects, objectId, null, null,
+                                'object', 'subject', 'predicate', graphId, fn)) {
+            return true;
+          }
+        }
+        else
+        // If nothing is given, iterate subjects and predicates first
+        if (this._someInIndex(graphContents.subjects, null, null, null,
+                              'subject', 'predicate', 'object', graphId, fn)) {
+          return true;
+        }
+      }
+    }
+    return false;
+  },
+
+  findGraphs: function (subject, predicate, object) {
+    var prefixes = this._prefixes;
+    return this.findGraphsByIRI(
+      expandPrefixedName(subject,   prefixes),
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(object,    prefixes)
+    );
+  },
+
+  findGraphsByIRI: function (subject, predicate, object) {
+    var out = [];
+    this.forGraphsByIRI(function (g) {
+      out.push(g);
+    }, subject, predicate, object);
+    return out;
+  },
+
+  forGraphs: function (fn, subject, predicate, object) {
+    var prefixes = this._prefixes;
+    this.forGraphsByIRI(
+      fn,
+      expandPrefixedName(subject,   prefixes),
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(object,    prefixes)
+    );
+  },
+
+  forGraphsByIRI: function (fn, subject, predicate, object) {
+    for (var graph in this._graphs) {
+      this.someByIRI(function (s, p, o, g) {
+        fn(g);
+        return true; // Halt iteration of some().
+      }, subject, predicate, object, graph);
+    }
+  },
+
+  findSubjects: function (predicate, object, graph) {
+    var prefixes = this._prefixes;
+    return this.findSubjectsByIRI(
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(object,    prefixes),
+      expandPrefixedName(graph,     prefixes)
+    );
+  },
+
+  findSubjectsByIRI: function (predicate, object, graph) {
+    var out = [];
+    this.forSubjectsByIRI(function (s) {
+      out.push(s);
+    }, predicate, object, graph);
+    return out;
+  },
+
+  forSubjects: function (fn, predicate, object, graph) {
+    var prefixes = this._prefixes;
+    this.forSubjectsByIRI(
+      fn,
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(object,    prefixes),
+      expandPrefixedName(graph,     prefixes)
+    );
+  },
+
+  forSubjectsByIRI: function (fn, predicate, object, graph) {
+    var seen = {}, ids = this._ids, graphs = {}, graphContents,
+        entityKeys = this._entities;
+
+    // Either loop over all graphs, or over just one selected graph.
+    if (!graph)
+      graphs = this._graphs;
+    else
+      graphs[graph] = this._graphs[graph];
+
+    // Translate IRIs to internal index keys.
+    // Optimization: if the entity doesn't exist, no triples with it exist.
+    if (predicate && !(predicate = ids[predicate])) return;
+    if (object    && !(object    = ids[object]))    return;
+
+    function handleResultsFn(id) {
+      if (!(id in seen)) {
+        seen[id] = null;
+        fn(entityKeys[id]);
+      }
+    }
+
+    for (graph in graphs) {
+      // Only if the specified graph contains triples, there can be results
+      if (graphContents = graphs[graph]) {
+        // We want to list all subjects.
+        // The three index choices are: SPO POS OSP.
+
+        // Choose optimal index based on which fields are wildcards.
+        if (predicate) {
+          if (object)
+            // If predicate and object are given, the POS index is best.
+            // Lookup p, lookup o, loop s.
+            this._lookupIndex0lookupIndex1loopIndex2(graphContents.predicates, predicate, object, handleResultsFn);
+          else
+            // If only predicate is given, the SPO index is best.
+            // Loop s, lookup p.
+            this._loopIndex0lookupIndex1(graphContents.subjects, predicate, handleResultsFn);
+        }
+        else if (object)
+          // If only object is given, the OSP index is best.
+          // Lookup o, loop s.
+          this._lookupIndex0loopIndex1(graphContents.objects, object, handleResultsFn);
+        else
+          // If no params given, iterate all the subjects.
+          this._loopIndex0(graphContents.subjects, handleResultsFn);
+      }
+    }
+  },
+
+  findPredicates: function (subject, object, graph) {
+    var prefixes = this._prefixes;
+    return this.findPredicatesByIRI(
+      expandPrefixedName(subject, prefixes),
+      expandPrefixedName(object,  prefixes),
+      expandPrefixedName(graph,   prefixes)
+    );
+  },
+
+  findPredicatesByIRI: function (subject, object, graph) {
+    var out = [];
+    this.forPredicatesByIRI(function (p) {
+      out.push(p);
+    }, subject, object, graph);
+    return out;
+  },
+
+  forPredicates: function (fn, subject, object, graph) {
+    var prefixes = this._prefixes;
+    this.forPredicatesByIRI(
+      fn,
+      expandPrefixedName(subject, prefixes),
+      expandPrefixedName(object,  prefixes),
+      expandPrefixedName(graph,   prefixes)
+    );
+  },
+
+  forPredicatesByIRI: function (fn, subject, object, graph) {
+    var seen = {}, ids = this._ids, graphs = this._graphs, graphContents,
+        entityKeys = this._entities;
+
+    // Either loop over all graphs, or over just one selected graph.
+    if (!graph)
+      graphs = this._graphs;
+    else
+      graphs[graph] = this._graphs[graph];
+
+    // Translate IRIs to internal index keys.
+    // Optimization: if the entity doesn't exist, no triples with it exist.
+    if (subject && !(subject = ids[subject])) return;
+    if (object  && !(object  = ids[object]))  return;
+
+    function handleResultsFn(id) {
+      if (!(id in seen)) {
+        seen[id] = null;
+        fn(entityKeys[id]);
+      }
+    }
+
+    for (graph in graphs) {
+      // Only if the specified graph contains triples, there can be results
+      if (graphContents = graphs[graph]) {
+        // We want to list all predicates.
+        // The three index choices are: SPO POS OSP.
+
+        // Choose optimal index based on which fields are wildcards.
+        if (subject) {
+          if (object)
+            // If subject and object are given, the OSP index is best.
+            // Lookup o, lookup s, loop p.
+            this._lookupIndex0lookupIndex1loopIndex2(graphContents.objects, subject, object, handleResultsFn);
+          else
+            // If only subject is given, the SPO index is best.
+            // lookup s, loop p.
+            this._lookupIndex0loopIndex1(graphContents.subjects, subject, handleResultsFn);
+        }
+        else if (object)
+          // If only object is given, the POS index is best.
+          // Loop p, lookup o.
+          this._loopIndex0lookupIndex1(graphContents.predicates, object, handleResultsFn);
+        else
+          // If no params given, iterate all the predicates.
+          this._loopIndex0(graphContents.predicates, handleResultsFn);
+      }
+    }
+  },
+
+  findObjects: function (subject, predicate, graph) {
+    var prefixes = this._prefixes;
+    return this.findObjectsByIRI(
+      expandPrefixedName(subject,   prefixes),
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(graph,     prefixes)
+    );
+  },
+
+  findObjectsByIRI: function (subject, predicate, graph) {
+    var out = [];
+    this.forObjectsByIRI(function (o) {
+      out.push(o);
+    }, subject, predicate, graph);
+    return out;
+  },
+
+  forObject: function (fn, subject, predicate, graph) {
+    var prefixes = this._prefixes;
+    this.forObjectsByIRI(
+      fn,
+      expandPrefixedName(subject,   prefixes),
+      expandPrefixedName(predicate, prefixes),
+      expandPrefixedName(graph,     prefixes)
+    );
+  },
+
+  forObjectsByIRI: function (fn, subject, predicate, graph) {
+    var seen = {}, ids = this._ids, graphs = this._graphs, graphContents,
+        entityKeys = this._entities;
+
+    // Either loop over all graphs, or over just one selected graph.
+    if (!graph)
+      graphs = this._graphs;
+    else
+      graphs[graph] = this._graphs[graph];
+
+    // Translate IRIs to internal index keys.
+    // Optimization: if the entity doesn't exist, no triples with it exist.
+    if (subject   && !(subject   = ids[subject]))   return;
+    if (predicate && !(predicate = ids[predicate])) return;
+
+    function handleResultsFn(id) {
+      if (!(id in seen)) {
+        seen[id] = null;
+        fn(entityKeys[id]);
+      }
+    }
+
+    for (graph in graphs) {
+      // Only if the specified graph contains triples, there can be results
+      if (graphContents = graphs[graph]) {
+        // We want to list all predicates.
+        // The three index choices are: SPO POS OSP.
+
+        // Choose optimal index based on which fields are wildcards.
+        if (subject) {
+          if (predicate)
+            // If subject and predicate are given, the SPO index is best.
+            // Lookup s, lookup p, loop o.
+            this._lookupIndex0lookupIndex1loopIndex2(graphContents.subjects, subject, predicate, handleResultsFn);
+          else
+            // If only subject is given, the OSP index is best.
+            // loop o, lookup s.
+            this._loopIndex0lookupIndex1(graphContents.objects, subject, handleResultsFn);
+        }
+        else if (predicate)
+          // If only predicate is given, the POS index is best.
+          // Looukp p, loop o.
+          this._lookupIndex0loopIndex1(graphContents.predicates, predicate, handleResultsFn);
+        else
+          // If no params given, iterate all the objects.
+          this._loopIndex0(graphContents.objects, handleResultsFn);
+      }
+    }
+  },
+
 
   // ### `createBlankNode` creates a new blank node, returning its name.
   createBlankNode: function (suggestedName) {