--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -133,7 +133,8 @@ N3Store.prototype = {
   _someInIndex: function (index0, key0, key1, key2, name0, name1, name2, graph, fn) {
     var tmp, index1, index2, varCount = !key0 + !key1 + !key2,
         // depending on the number of variables, keys or reverse index are faster
-        entityKeys = varCount > 1 ? Object.keys(this._ids) : this._entities;
+        entityKeys = varCount > 1 ? Object.keys(this._ids) : this._entities,
+        result = { subject: '', predicate: '', object: '', graph: graph };
 
     // If a key is specified, use only that part of index 0.
     if (key0) (tmp = index0, index0 = {})[key0] = tmp[key0];
@@ -151,7 +152,6 @@ N3Store.prototype = {
             var values = key2 ? (key2 in index2 ? [key2] : []) : Object.keys(index2);
             // Create triples for all items found in index 2.
             for (var l = values.length - 1; l >= 0; l--) {
-              var result = { subject: '', predicate: '', object: '', graph: graph };
               result[name0] = entity0;
               result[name1] = entity1;
               result[name2] = entityKeys[values[l]];