--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -139,21 +139,19 @@ N3Store.prototype = {
     // If a key is specified, use only that part of index 0.
     if (key0) (tmp = index0, index0 = {})[key0] = tmp[key0];
     for (var value0 in index0) {
-      var entity0 = entityKeys[value0];
+      result[name0] = entityKeys[value0];
 
       if (index1 = index0[value0]) {
         // If a key is specified, use only that part of index 1.
         if (key1) (tmp = index1, index1 = {})[key1] = tmp[key1];
         for (var value1 in index1) {
-          var entity1 = entityKeys[value1];
+          result[name1] = entityKeys[value1];
 
           if (index2 = index1[value1]) {
             // If a key is specified, use only that part of index 2, if it exists.
             var values = key2 ? (key2 in index2 ? [key2] : []) : Object.keys(index2);
             // Create triples for all items found in index 2.
             for (var l = values.length - 1; l >= 0; l--) {
-              result[name0] = entity0;
-              result[name1] = entity1;
               result[name2] = entityKeys[values[l]];
               if (fn(result.subject, result.predicate, result.object, result.graph)) {
                 return true;