--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -164,39 +164,31 @@ N3Store.prototype = {
     return false;
   },
 
-  _lookupIndex0lookupIndex1loopIndex2: function (index0, key0, key1, fn) {
+  _index2KeysGivenKey0And1: function (index0, key0, key1, fn) {
     var index1, index2, key2;
-    // Lookup.
     if (!(index1 = index0[key0])) return;
-    // Lookup.
     if (!(index2 = index1[key1])) return;
-    // Loop.
     for (key2 in index2)
       fn(key2);
   },
 
-  _lookupIndex0loopIndex1: function (index0, key0, fn) {
+  _index1KeysGivenKey0: function (index0, key0, fn) {
     var index1, key1;
-    // Lookup.
     if (!(index1 = index0[key0])) return;
-    // Loop.
     for (key1 in index1)
       fn(key1);
   },
 
-  _loopIndex0lookupIndex1: function (index0, key1, fn) {
+  _index0KeysGivenKey1: function (index0, key1, fn) {
     var key0, index1;
-    // Loop.
     for (key0 in index0) {
       index1 = index0[key0];
-      // Lookup.
       if (index1[key1])
         fn(key0);
     }
   },
 
-  _loopIndex0: function (index0, fn) {
-    // Loop.
+  _index0Keys: function (index0, fn) {
     for (var key0 in index0)
       fn(key0);
   },
@@ -637,19 +629,19 @@ N3Store.prototype = {
           if (object)
             // If predicate and object are given, the POS index is best.
             // Lookup p, lookup o, loop s.
-            this._lookupIndex0lookupIndex1loopIndex2(graphContents.predicates, predicate, object, handleResultsFn);
+            this._index2KeysGivenKey0And1(graphContents.predicates, predicate, object, handleResultsFn);
           else
             // If only predicate is given, the SPO index is best.
             // Loop s, lookup p.
-            this._loopIndex0lookupIndex1(graphContents.subjects, predicate, handleResultsFn);
+            this._index0KeysGivenKey1(graphContents.subjects, predicate, handleResultsFn);
         }
         else if (object)
           // If only object is given, the OSP index is best.
           // Lookup o, loop s.
-          this._lookupIndex0loopIndex1(graphContents.objects, object, handleResultsFn);
+          this._index1KeysGivenKey0(graphContents.objects, object, handleResultsFn);
         else
           // If no params given, iterate all the subjects.
-          this._loopIndex0(graphContents.subjects, handleResultsFn);
+          this._index0Keys(graphContents.subjects, handleResultsFn);
       }
     }
   },
@@ -714,19 +706,19 @@ N3Store.prototype = {
           if (object)
             // If subject and object are given, the OSP index is best.
             // Lookup o, lookup s, loop p.
-            this._lookupIndex0lookupIndex1loopIndex2(graphContents.objects, subject, object, handleResultsFn);
+            this._index2KeysGivenKey0And1(graphContents.objects, subject, object, handleResultsFn);
           else
             // If only subject is given, the SPO index is best.
             // lookup s, loop p.
-            this._lookupIndex0loopIndex1(graphContents.subjects, subject, handleResultsFn);
+            this._index1KeysGivenKey0(graphContents.subjects, subject, handleResultsFn);
         }
         else if (object)
           // If only object is given, the POS index is best.
           // Loop p, lookup o.
-          this._loopIndex0lookupIndex1(graphContents.predicates, object, handleResultsFn);
+          this._index0KeysGivenKey1(graphContents.predicates, object, handleResultsFn);
         else
           // If no params given, iterate all the predicates.
-          this._loopIndex0(graphContents.predicates, handleResultsFn);
+          this._index0Keys(graphContents.predicates, handleResultsFn);
       }
     }
   },
@@ -791,19 +783,19 @@ N3Store.prototype = {
           if (predicate)
             // If subject and predicate are given, the SPO index is best.
             // Lookup s, lookup p, loop o.
-            this._lookupIndex0lookupIndex1loopIndex2(graphContents.subjects, subject, predicate, handleResultsFn);
+            this._index2KeysGivenKey0And1(graphContents.subjects, subject, predicate, handleResultsFn);
           else
             // If only subject is given, the OSP index is best.
             // loop o, lookup s.
-            this._loopIndex0lookupIndex1(graphContents.objects, subject, handleResultsFn);
+            this._index0KeysGivenKey1(graphContents.objects, subject, handleResultsFn);
         }
         else if (predicate)
           // If only predicate is given, the POS index is best.
           // Looukp p, loop o.
-          this._lookupIndex0loopIndex1(graphContents.predicates, predicate, handleResultsFn);
+          this._index1KeysGivenKey0(graphContents.predicates, predicate, handleResultsFn);
         else
           // If no params given, iterate all the objects.
-          this._loopIndex0(graphContents.objects, handleResultsFn);
+          this._index0Keys(graphContents.objects, handleResultsFn);
       }
     }
   },