--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -674,7 +674,7 @@ N3Store.prototype = {
   },
 
   forPredicatesByIRI: function (fn, subject, object, graph) {
-    var seen = {}, ids = this._ids, graphs = this._graphs, graphContents,
+    var seen = {}, ids = this._ids, graphs = {}, graphContents,
         entityKeys = this._entities;
 
     // Either loop over all graphs, or over just one selected graph.
@@ -751,7 +751,7 @@ N3Store.prototype = {
   },
 
   forObjectsByIRI: function (fn, subject, predicate, graph) {
-    var seen = {}, ids = this._ids, graphs = this._graphs, graphContents,
+    var seen = {}, ids = this._ids, graphs = {}, graphContents,
         entityKeys = this._entities;
 
     // Either loop over all graphs, or over just one selected graph.