--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -133,27 +133,29 @@ N3Store.prototype = {
   _someInIndex: function (index0, key0, key1, key2, name0, name1, name2, graph, fn) {
     var tmp, index1, index2, varCount = !key0 + !key1 + !key2,
         // depending on the number of variables, keys or reverse index are faster
-        entityKeys = varCount > 1 ? Object.keys(this._ids) : this._entities,
-        result = { subject: '', predicate: '', object: '', graph: graph };
+        entityKeys = varCount > 1 ? Object.keys(this._ids) : this._entities;
 
     // If a key is specified, use only that part of index 0.
     if (key0) (tmp = index0, index0 = {})[key0] = tmp[key0];
     for (var value0 in index0) {
-      result[name0] = entityKeys[value0];
+      var entity0 = entityKeys[value0];
 
       if (index1 = index0[value0]) {
         // If a key is specified, use only that part of index 1.
         if (key1) (tmp = index1, index1 = {})[key1] = tmp[key1];
         for (var value1 in index1) {
-          result[name1] = entityKeys[value1];
+          var entity1 = entityKeys[value1];
 
           if (index2 = index1[value1]) {
             // If a key is specified, use only that part of index 2, if it exists.
             var values = key2 ? (key2 in index2 ? [key2] : []) : Object.keys(index2);
             // Create triples for all items found in index 2.
             for (var l = values.length - 1; l >= 0; l--) {
+              var result = { subject: '', predicate: '', object: '', graph: graph };
+              result[name0] = entity0;
+              result[name1] = entity1;
               result[name2] = entityKeys[values[l]];
-              if (fn(result.subject, result.predicate, result.object, result.graph)) {
+              if (fn(result)) {
                 return true;
               }
             }
@@ -437,8 +439,8 @@ N3Store.prototype = {
   },
 
   forEachByIRI: function (fn, subject, predicate, object, graph) {
-    this.someByIRI(function (s, p, o, g) {
-      fn(s, p, o, g);
+    this.someByIRI(function (quad) {
+      fn(quad);
       return false;
     }, subject, predicate, object, graph);
   },
@@ -455,8 +457,8 @@ N3Store.prototype = {
   },
 
   everyByIRI: function (fn, subject, predicate, object, graph) {
-    return !this.someByIRI(function (s, p, o, g) {
-      return !fn(s, p, o, g);
+    return !this.someByIRI(function (quad) {
+      return !fn(quad);
     }, subject, predicate, object, graph);
   },
 
@@ -562,8 +564,8 @@ N3Store.prototype = {
 
   forGraphsByIRI: function (fn, subject, predicate, object) {
     for (var graph in this._graphs) {
-      this.someByIRI(function (s, p, o, g) {
-        fn(g);
+      this.someByIRI(function (quad) {
+        fn(quad.graph);
         return true; // Halt iteration of some().
       }, subject, predicate, object, graph);
     }