--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -742,7 +742,7 @@ N3Store.prototype = {
     return out;
   },
 
-  forObject: function (fn, subject, predicate, graph) {
+  forObjects: function (fn, subject, predicate, graph) {
     var prefixes = this._prefixes;
     this.forObjectsByIRI(
       fn,