--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -266,7 +266,9 @@ N3Store.prototype = {
 
     // It turns out that varCount is actually a count of null parameters.
     // We always have a null parameter count of 2.
-    var entityKeys = Object.keys(this._ids);
+    // var entityKeys = Object.keys(this._ids); // Fastest - according to above rule
+
+    var entityKeys = this._entities; // But I'm actually seeing this as a fraction faster?
 
     // console.log('given key0, varCount ', varCount);