--- a/lib/N3Store.js
+++ b/lib/N3Store.js
@@ -272,6 +272,8 @@ N3Store.prototype = {
 
     // console.log('given key0, varCount ', varCount);
 
+    // TODO(js) This is the method that /appears/ to be performing more slowly then before.
+
     // Lookup
     if (key0 in index0) {
       var index1 = index0[key0];