This is the result of an experiment. It's not really a pull request, although you're welcome to take the code if you wish :)

So... I wrote the 'four functions' version of `some`, and hooked up `find` to use `some`.

My benchmarks for current `master` are typically:

```
N3Store performance test
- Adding 16777216 triples to the default graph: 18528.181ms
* Memory usage for triples: 709MB
- Finding all 16777216 triples in the default graph 65536 times (0 variables): 88729.043ms
- Finding all 16777216 triples in the default graph 131072 times (1 variable): 5624.881ms
- Finding all 16777216 triples in the default graph 196608 times (2 variables): 5504.780ms

- Adding 16777216 quads: 29252.079ms
* Memory usage for quads: 673MB
- Finding all 16777216 quads 1048576 times: 25739.851ms
```

And the benchmarks for this branch can be in the region of:

```
$ node N3Store-perf.js
N3Store performance test
- Adding 16777216 triples to the default graph: 21008.823ms
* Memory usage for triples: 719MB
- Finding all 16777216 triples in the default graph 65536 times (0 variables): 39437.696ms
- Finding all 16777216 triples in the default graph 131072 times (1 variable): 4961.082ms
- Finding all 16777216 triples in the default graph 196608 times (2 variables): 7438.038ms

- Adding 16777216 quads: 27909.561ms
* Memory usage for quads: 799MB
- Finding all 16777216 quads 1048576 times: 16721.452ms
```

Hmmm. Curious indeed.

The problem function in this new experimental code is `_indexSomeGivenKey0` (called for '2 variables', above). Regarding the voodoo trick: my benchmarks seem to show that it doesn't matter whether I use the `Object.keys` or `this._entities` here.

- I get very inconsistent benchmark times.
- Partly I think this may be due to my laptop turbo-boosting or throttling. Don't know. It's not the best machine to benchmark on really because of this.
- Partly I think it may be because the benchmark is actually timing a large chunk of work compared to my previous optimisation experience?
- Partly I think it may be because it's an interpreted language. But I am biased, being much more familiar with compiled languages.
- It's so ridiculously fast on the other methods/permutations that I'm beginning to think we've definitely hit a weird edge case in V8 (take a look at `_indexSomeGivenKey0`... it couldn't do any less - except for not having a callback if we were doing `find`, I guess. But seriously?) Maybe the Chrome team would be interested?
- Apologies in advance for all the commented-out dead code
- Some of the code is plain ugly - and no apologies could really cover that. 

Anyhow, make of it what you will. (Good luck)

— Have you perhaps considered implementing the index lookups in asm.js? <gets coat and runs>