@jimsmart 2017-01-07
Travis hates my lack of adherence to linting - but eslint hated my code :(