@jimsmart 2017-01-08
Having given this a little thought now: if anything, perhaps it's worth keeping just `_indexSomeWhenGivenAllKeys` from this branch - that seems to be a very big win for small additional code / code changes. 

Performance vs code duplication / readability / maintenance, often a tricky call.