@jimsmart 2017-01-09
One problem with benchmarks is that they can make you focus on the wrong things...

Looking through my code here, I don't have even a single call to `some` / `find` / etc with _all_ of the parameters given. 

I don't believe it's a common use case - it's every other case we want optimising - I don't think that case is worthwhile optimising for. 

— Even if it does look good on the benchmark.