@RubenVerborgh 2017-04-15
Indeed—to be looked at when we have better real-world benchmarking scenarios.